chore(core): Move Rime class into core package
