+++ app/src/main/java/com/osfans/trime/core/Rime.java
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.osfans.trime;
+package com.osfans.trime.core;
 
 import android.content.Context;
 import android.text.TextUtils;
