+++ app/src/main/java/com/osfans/trime/ime/core/EditorInstance.kt
@@ -8,7 +8,7 @@ import android.view.KeyEvent
 import android.view.inputmethod.EditorInfo
 import android.view.inputmethod.ExtractedTextRequest
 import android.view.inputmethod.InputConnection
-import com.osfans.trime.Rime
+import com.osfans.trime.core.Rime
 import com.osfans.trime.ime.enums.InlineModeType
 import com.osfans.trime.ime.text.TextInputManager
 import timber.log.Timber
