+++ app/src/main/java/com/osfans/trime/ime/text/ScrollView.java
@@ -9,7 +9,7 @@ import android.view.View;
 import android.view.animation.TranslateAnimation;
 import android.widget.HorizontalScrollView;
 import androidx.annotation.NonNull;
-import com.osfans.trime.Rime;
+import com.osfans.trime.core.Rime;
 import timber.log.Timber;
 
 public class ScrollView extends HorizontalScrollView {
