+++ app/src/main/java/com/osfans/trime/settings/components/SchemaPickerDialog.kt
@@ -7,7 +7,7 @@ import android.os.Build
 import android.view.WindowManager
 import androidx.appcompat.app.AlertDialog
 import com.osfans.trime.R
-import com.osfans.trime.Rime
+import com.osfans.trime.core.Rime
 import com.osfans.trime.ime.core.Trime
 import com.osfans.trime.setup.Config
 import com.osfans.trime.util.RimeUtils
